+++ src/main/java/org/kdb/inside/brains/core/KdbQueryListener.java
@@ -4,6 +4,6 @@ public interface KdbQueryListener {
     default void queryStarted(InstanceConnection connection, KdbQuery query) {
     }
 
-    default void queryFinished(InstanceConnection connection, KdbQuery query) {
+    default void queryFinished(InstanceConnection connection, KdbQuery query, KdbResult result) {
     }
 }